fix:correção de apontamentos do PR
